move Component.js code to index.js
